Fix amd64 warning
git-svn-id: svn://coreboot.org/openbios/openbios-devel@393 f158a5a8-5612-0410-a976-696ce0be7e32

+++ include/amd64/types.h
@@ -18,8 +18,8 @@
 
 /* cell based types */
 
-typedef int64_t       cell;
-typedef uint64_t     ucell;
+typedef long long     cell;
+typedef unsigned long long ucell;
 typedef __int128_t   dcell;
 typedef __uint128_t ducell;
